file is loaded successfully into pandas. Now lets make sure the data is being labelled and processed correctly
