+++ v2/requirements.txt
@@ -1,268 +1,40 @@
-absl-py==0.7.1
-altair==3.2.0
-appdirs==1.4.3
-appnope==0.1.0
-argh==0.26.2
-arrow==0.15.5
-arviz==0.5.1
-astor==0.8.0
-astroid==2.2.5
-atomicwrites==1.3.0
-attrs==18.2.0
-autopep8==1.5.2
-backcall==0.1.0
-base58==1.0.3
-beautifulsoup4==4.7.1
-binaryornot==0.4.4
-bleach==3.1.0
-blinker==1.4
-blis==0.4.1
-bokeh==1.0.4
-boto3==1.13.3
-botocore==1.16.3
-branca==0.3.1
-brewer2mpl==1.4.1
-bs4==0.0.1
-cached-property==1.5.1
-cachetools==3.1.0
-catalogue==0.2.0
-certifi==2019.3.9
-cffi==1.14.0
-cftime==1.0.3.4
-chardet==3.0.4
-chart-studio==1.0.0
+boto3==1.13.6
+botocore==1.16.6
+certifi==2019.11.28
 Click==7.0
-click-plugins==1.1.1
-cligj==0.5.0
 coloredlogs==10.0
-cookiecutter==1.7.0
-cryptography==2.8
-cycler==0.10.0
-cymem==2.0.3
 dash==1.9.0
 dash-core-components==1.8.0
 dash-dangerously-set-inner-html==0.0.2
-dash-daq==0.3.1
 dash-flow-example==0.0.5
 dash-html-components==1.0.2
 dash-renderer==1.2.4
 dash-table==4.6.0
-dash.ly==0.17.3
-DateTime==4.3
-decorator==4.3.2
-defusedxml==0.5.0
 docutils==0.15.2
-dstack==0.2.0
-en-core-web-sm==2.2.5
 entrypoints==0.3
-enum-compat==0.0.3
-et-xmlfile==1.0.1
-Fiona==1.8.6
 fire==0.2.1
 flake8==3.7.9
 Flask==1.1.1
-Flask-Caching==1.8.0
 Flask-Compress==1.4.0
-Flask-SeaSurf==0.2.2
-floweaver==2.0.0a4
-folium==0.8.3
 future==0.18.2
-fuzzywuzzy==0.17.0
-gast==0.2.2
-geopandas==0.5.1
-ggplot==0.11.5
-gitdb2==2.0.5
-githubcommit==0.1.0
-gitlab==1.0.2
-GitPython==2.1.11
-google-api-core==1.7.0
-google-api-python-client==1.7.9
-google-auth==1.6.2
-google-auth-httplib2==0.0.3
-google-auth-oauthlib==0.4.0
-google-cloud-core==0.29.1
-google-cloud-storage==1.13.2
-google-pasta==0.1.7
-google-resumable-media==0.3.2
-googleads==21.0.0
-googleapis-common-protos==1.5.6
-graphviz==0.10.1
-grpcio==1.22.0
-gunicorn==20.0.4
-h5py==2.9.0
-html5lib==1.0.1
-httplib2==0.13.0
 humanfriendly==6.0
-idna==2.8
-imbalanced-learn==0.4.3
-imblearn==0.0
-importlib-metadata==1.3.0
-inflect==4.0.0
-ipykernel==5.1.0
-ipysankeywidget==0.2.5
-ipython==7.2.0
-ipython-genutils==0.2.0
-ipywidgets==7.4.2
-isodate==0.6.0
-isort==4.3.21
 itsdangerous==1.1.0
-jdcal==1.4
-jedi==0.13.2
-Jinja2==2.10.3
-Jinja2-Ext==0.1
-jinja2-time==0.2.0
+Jinja2==2.11.1
 jmespath==0.9.5
-joblib==0.14.1
-jsonschema==2.6.0
-jupyter==1.0.0
-jupyter-client==5.2.4
-jupyter-console==6.0.0
-jupyter-contrib-core==0.3.3
-jupyter-contrib-nbextensions==0.5.1
-jupyter-core==4.4.0
-jupyter-highlight-selected-word==0.2.0
-jupyter-latex-envs==1.4.6
-jupyter-nbextensions-configurator==0.4.1
-Keras==2.2.4
-Keras-Applications==1.0.8
-Keras-Preprocessing==1.1.0
-kiwisolver==1.2.0
-lazy-object-proxy==1.4.3
-List==1.3.0
-lxml==4.3.0
-Markdown==3.1.1
 MarkupSafe==1.1.1
-matplotlib==3.2.1
 mccabe==0.6.1
-merf==0.3
-missingno==0.4.1
-mistune==0.8.4
-mlens==0.2.3
-mlxtend==0.17.0
-mock==2.0.0
-more-itertools==7.0.0
-mpld3==0.3
-munch==2.3.2
-murmurhash==1.0.2
-nbconvert==5.4.0
-nbformat==4.4.0
-nbresuse==0.3.3
-netCDF4==1.5.2
-networkx==1.11
-nltk==3.4.5
-notebook==5.7.4
+mock==4.0.1
 numpy==1.18.4
-oauth2client==4.1.3
-oauthlib==3.0.1
-openpyxl==2.6.0
-packaging==19.0
-palettable==3.1.1
 pandas==1.0.3
-pandocfilters==1.4.2
-parso==0.3.2
-pathtools==0.1.2
-patsy==0.5.1
-pbr==5.1.3
-percy==2.0.2
-pexpect==4.6.0
-pickleshare==0.7.5
-Pillow==6.2.1
-plac==1.1.3
-plotly==4.4.1
-pluggy==0.9.0
-poyo==0.5.0
-preshed==3.0.2
-probscale==0.2.3
-prometheus-client==0.5.0
-prompt-toolkit==2.0.8
-protobuf==3.6.1
-psutil==5.6.1
-ptyprocess==0.6.0
-py==1.8.0
-pyasn1==0.4.5
-pyasn1-modules==0.2.4
+plotly==4.5.0
 pycodestyle==2.5.0
-pycparser==2.19
-pydotplus==2.0.2
 pyflakes==2.1.1
-Pygments==2.3.1
-pylint==2.3.1
-pymc3==3.7
-pyOpenSSL==19.1.0
-pyparsing==2.4.7
-PyPDF2==1.26.0
-pyperclip==1.7.0
-pyproj==2.2.1
-pytest==4.4.0
-pytest-mock==2.0.0
-pytest-sugar==0.9.2
 python-dateutil==2.8.1
-python-Levenshtein==0.12.0
 pytz==2020.1
-PyYAML==5.1.2
-pyzmq==17.1.2
-qtconsole==4.4.3
-redis==3.3.11
-requests==2.21.0
-requests-oauthlib==1.2.0
-requests-toolbelt==0.9.1
-researchpy==0.1.7
+PyYAML==5.3
 retrying==1.3.3
-rq==1.2.0
-rsa==4.0
 s3transfer==0.3.3
-scikit-learn==0.22.2.post1
-scipy==1.4.1
-seaborn==0.9.0
-selenium==3.141.0
-Send2Trash==1.5.0
-Shapely==1.6.4.post2
 six==1.14.0
-sklearn==0.0
-smmap2==2.0.5
-soupsieve==1.8
-spacy==2.2.3
-SQLAlchemy==1.3.1
-srsly==0.2.0
-statsmodels==0.10.1
-streamlit==0.50.2
-tensorboard==1.14.0
-tensorflow==1.14.0
-tensorflow-estimator==1.14.0
 termcolor==1.1.0
-terminado==0.8.1
-testfixtures==6.7.0
-testpath==0.4.2
-textblob==0.15.3
-Theano==1.0.4
-thinc==7.3.1
-toml==0.10.0
-toolz==0.10.0
-tornado==5.1.1
-tqdm==4.36.1
-traitlets==4.3.2
-typed-ast==1.4.1
-tzlocal==2.0.0
-uritemplate==3.0.0
 urllib3==1.25.9
-validators==0.14.0
-virtualenv==16.3.0
-waitress==1.4.3
-wasabi==0.4.2
-watchdog==0.9.0
-watson-developer-cloud==2.7.0
-wcwidth==0.1.7
-webencodings==0.5.1
-websocket-client==0.48.0
-Werkzeug==0.16.0
-whichcraft==0.6.1
-widgetsnbextension==3.4.2
-wrapt==1.11.2
-xarray==0.13.0
-xgboost==0.90
-xlrd==1.2.0
-xlwt==1.3.0
-xmltodict==0.12.0
-zeep==3.4.0
-zipp==0.6.0
-zope.interface==4.6.0
+Werkzeug==1.0.0
